SF: Fix DisplayTransactionTest
SetupNewDisplayDeviceInternalTest.createHwcVirtualDisplay was failing
after presubmit unit tests were enabled on pi-dev-plus-aosp.

It turned out somehow the test had gotten out of sync with the
implementation, and the expectations did not match for one of the cases
tested.

This patch simply fixes the test to use the correct expectations for
what calls are made for a HWC supported virtual display when one is
created.

Bug: 118887786
Test: Presubmit passes

Change-Id: I366e51d81ae3cb048e7a287e4f859ac92c55a20f