Allow the same producer to gain a buffer twice.
This change makes the same producer gaining the same buffer the second
time a no-op, and prevent other clients of the buffer from gaining it.

Currently, EALREADY error will return if a producer gain a buffer which
is already gained by the producer. Problem: there is an actual use case
in passthrough processor that a producer may need to gain a buffer to
mimic a "release" of a buffer in the consumer point of view.

Test: buffer_hub-test
Bug: 118718713
Change-Id: If958c95165a5ad62f51c46b01fa43b576d3b20ae