Simplify process(..) in InputReader_test
The call to process(..) already includes a mapper as a parameter. An
InputMapper is always associated with exactly one InputDevice. As a
result, when the mapper processes a RawEvent, it really shouldn't care
about deviceId anymore.

For completeness, we populate the device id from the mapper's
getDeviceId() function.
This allows the calls to be simplified throughout the code.

Test: atest inputflinger_tests
Bug: none
Change-Id: I85824629044c13da8486c3c0e9fcf3550b957cd7