SF: Create a cicular buffer structure to store layer information.
This structure is going to hold information about history of layers. For now,
we keep information about layers inserted and their presentation timestamp.

Every container (unordered_map) holds information about layers presented at a given
frame. At the beginning of frame, we clear the map and add every layer that had an update
at that frame.

We expose a function to increment the counter of the structure, so that can be
incremented each time SurfaceFlinger wakes up.

see go/surface-flinger-scheduler for more info

Test: Adding a unit test for the simple structure.
Bug: 113612090
Change-Id: If120a15b5f9a0883e70b348dc8cca1b2e4f85f5d