Fix logic error in atrace when detecting available categories.
A category should be marked as available "only if all its required /sys/
files are writeable and if enabling the category will enable one or more
tracing tags or /sys/ files". The previous code didn't check that one or
more file was enabled, and therefore would always mark a category available
if all of their files were optional, even if none of those were
writeable.

Bug: 120621110
Test: Manual
Change-Id: Iee2984f7cb3591dcd0b3a376d95754d1abaea5be