Add vector of TouchVideoFrame to NotifyMotionArgs
NotifyMotionArgs is passed to InputListener. We add a vector of
TouchVideoFrame here to NotifyMotionArgs  in order to provide heatmap
information about touch. This heatmap information can later be used to
do further processing of touch. For example, it could be used to
determine whether a given touch stream represents a user intentionally
pressing harder on the screen.

Introduce a new class, TouchVideoFrame, that will represent a single
scan of the touch heatmap for a given touch event.

Test: integration tested by adding a debug log to the case where the
finger area is large. That means, the video frames are being passed
locally to the InputClassfier HAL. Although HAL is not being submitted
together with this change, the current change will make the subsequent
CLs simpler to review.
Bug: 111480215

Change-Id: I4f16c8b2bd726c6bd4bbd5d2b18d2536a2347bda