Add a watch for touch video devices to EventHub
EventHub will now be responsible for reading touch video devices in
addition to the input devices. We add the initial infrastructure for
that here.

We keep track of the watch descriptor inside inotify, and check which
event was generated.

All of the video devices go directly into /dev. For the heatmap-specific
video devices, the prefix will be "v4l-touch". For example, a
touchscreen device on C1 is currently "/dev/v4l-touch22". Although we
don't anticipate having pluggable video devices which report touch
heatmaps, we should still account for that possibility here for
completeness. Every time a device appears in /dev after the initial
scan, this inotify will trigger the check for whether it is a v4l2
touchscreen node.

Currently, we just add a log for when the device appears, without taking
any action. In the future commits, we will create touch video devices
for this node, and associate them with input devices.

Bug: 62940136
Test: interaction with booted device
Change-Id: I99bf29c3786ec4b21fed1ddf2e1e87f5a53c172a