Have `otapreopt_chroot` mount APEX packages in the chroot dir.
We need at least the Android Runtime APEX to be able to run `dex2oat`
in `otapreopt`. For simplicity, reuse `apexd`'s logic and mount all
the APEX packages found on the system partition in /postinstall/apex.

This is a temporary change. The final design will integrate A/B OTA
updates and `apexd`.

Test: m otapreopt_chroot
Test: A/B OTA update test (asit/dexoptota/self_full).
Bug: 113373927
Bug: 120796514
Change-Id: Ib7d37fa889d368651ca2b02f49e8fffea80dd7bf