Associate touch video devices with input devices
In earlier commits, we added touch video devices, but we did not add the
relationship between the input device and the video device. Add the
relationship here.

The matching between input devices and video devices is done by name
(for now). This was determined after reviewing the api for v4l2 devices,
which do not contain any obvious matching techniques, and do not have
.uniq values. This logic may be changed later.
The infrastructure being set up here, however, will remain as is.

Test: dumpsys input of a booted device, and interact with the device by
touch.
Bug: 62940136

Change-Id: I116e6ece8086359cd6546db7af6b2db707fca1e3