Input: Update input window info when the layer receives its first buffer
Currently input window info is populated when a surface flinger transaction is applied. When
populating input window info for a layer that is yet to receive its first buffer, the window
will be invisible and not receive any input until another transaction is applied. This can
result in input being dispatched to a window beneath the target window.

To fix, populate the input window info when we latch a new buffer as well.

Test: sf unit tests+go/wm-smoke
Bug: 120839715
Change-Id: I42f9dc3a64a7af2f3359f0dfeefb57f7a3e75306