Add MotionClassification to NotifyMotionArgs
This classification will be set in the InputClassifier stage and will
get passed on to MotionEvent. It will be consumed in user-facing API
MotionEvent.getClassification() that will provide an IntDef
classification.

The ultimate user will be GestureDetector, View, and other similar APIs
that provide information about long press. The goal is to provide means
for accelerating the long-press gesture.

Bug: 62940136
Test: Integration tested in upper layers on MotionEvent.
Change-Id: I66e062c21c6a860a3abb16e265e37b18ebef1bcb