SurfaceFlinger: Only allow update of InputWindowInfo with permission.
This was intended to be in the original CL, but it looks like I never added it or
lost it at some point. This is important as otherwise clients could use modality flags,
global focus, etc...to escape their constraints. The model is the WM can "bless" a surface for input
with given flags (e.g. non-modal) and then the client can manipulate the geometry if it
happens to have the surfacecontrol, but within the constraints of the hierarchy.

Bug: 111440400
Test: Builds
Change-Id: Ifea0882bb26f9791aa2736895c900a05a02ba7cb