Revert "Respect buffer source crop in client composition."
This reverts commit e0451f753d51e33afde34c7b6837c5824169c59f.

Reason for revert: this doesn't quite work in split screen when the device is rotated due to some buffer size mismatch. I have a forward fix in mind, but I don't believe it requires some of the interface changes so reverting first even though no one's caught this yet :x.

Change-Id: I1a0a0d028beb13f810ac16eb8d112d7c89b1cfca