Changed ContainerLayer's isVisible method to always return false.
ContainerLayers should always be considered not visible so they don't get
calculated when computing visible regions. However, isVisible is also used when
sending window info to InputManager about which windows can be touched.
ContainerLayers should still be considered visible to InputManager. The
function canReceiveInput was added so it can return a different value than
isVisible which only means visible to the screen.

Test: No longer composer ContainerLayers
Bug: 123686354
Change-Id: Ib39952e89ffd4ea9618b6027e7da96ef175de058