SF: Fixup call expectations in libsurfaceflinger_unittest
This is done to eliminate a flood of warning messages about unexpected
calls when the test is run directly (not visible when run with atest)

Most of the warnings had to do with changes to the implementation to
query HWC capabilities when a display is connected, and other queries to set
up the framebuffer surface afterwards.

With these warnings gone, it is much easier to see other unexpected
calls that might actually be important.

Test: atest libsurfaceflinger_unittest
Bug: None
Change-Id: Ia386daa9dcbf9cfbcd3e12be3b63f6e05fdcfe5e