SurfaceFlinger: Ensure reparent triggers visible region calculation.
Following the recent CL there are still more errors. First is that we may
not set the transaction flags, leading to SurfaceFlinger failing to call
doTransaction. We fix this by calling setTransactionFlags from add/removeChild.
The second problem is that if mCurrentState.modified were not true, do transaction
would return early and fail to produce eVisibleRegions. We fix this with a refactoring
in doTransaction.

Test: Manual
Bug: 123333167
Change-Id: Iccd25244b6cb38a1fd90dc5165820d1100f40f4a