Fix style violation according to go/droidcppstyle
Local variable should be lowerCamelCase.

This change leave the buffer_state fence_state,
active_clients_bit_mask and queue_index in MetadataHeader unchanged
because there are too many instances of them that need to be changed in
both new bufferhub system and the old pdx-backed bufferhubd system.
I will do a separate change for that.

Test: m, mma
Bug: 68273829
Change-Id: Ia91cf7debe01a402f61171f069988e0ac13cec6a